dtls: use functions to extract numbers
These defines did not had a good type check in it and just casted the
parameters. This caused problems on 16 bit CPUs as there int is 16 bits
and the default value for numbers. Now the compiler is able to change
the given integer to the needed size. This also fixes 2 bugs where the
wrong pointer was used in the code.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>