dtls: add DTLS prefix in front of header define
This define is there to prevent this actual header from being included
more than one time, but this should be unique in all header files used
in all code tinydtls is compiled with. Add a prefix to prevent
collisions with other code we do not control. This is one part to make
tinydtls work better in combination with libcoap.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>