dtls: add dtls prefix to prng functions
This should prevent collisions with other code compiled with tinydtls
into one binary. The include from crypto.h was removed to prevent this
file from being included in user applications using tinydtls.

The functions and defines in prng.h got some changes to make the easier
to read. This is one part to make tinydtls work better in combination
with libcoap.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>