dtls: use dtls_send() for sending a handshake message
This changes dtls_send() to dtls_send_multi() which is able to take an
array of buffers to send and not just one. This is now also used in
dtls_send_handshake_msg_hash() to send the handshake messages, because
dtls_send() also handles retransmission and handshake messages should
be retransmitted when needed.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>