dtls: mark the dtls header structs packed
For some processor architectures like the MSP430X the compiler, in this
case gcc, optimizes the structs to align to full words length. This is
done because unaligned memory access is costly or not possible on some
architectures. GCC then adds some padding into structs which places
some members at different offsets and increases the size. This makes
the record header on the MSP430X 1 byte bigger for example, which is
wrong.

This patch adds the __packed__ attribute to the structs which are
representing a header defined in the DTLS RFC where the compiler should
not add any pads into the struct.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>