dtls: client should expect CCS from server
Without this patch the client expected a Finish message after it
received the ServerHelloDone message from the server. This worked
because the ChangeCipherSpec message was send in the old epoch and the
client was already in the new epoch and ignored it. Now this message
goes through and the client should handle it.

This uniforms the server and client states.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>