dtls: reorder struct netq_t to remove padding
The compiler adds paddings into structs if they do not align for the
bitsize of the CPU and this increases the size of some structures.
This is needed for some CPU, but it is better to order the members in a
strcut in a way that padding is not needed.

This reduces the bss segment size on msp430X:

old:
   text    data     bss     dec     hex filename
    264      12     892    1168     490 obj_wismote/netq.o

new:
   text    data     bss     dec     hex filename
    264      12     884    1160     488 obj_wismote/netq.o

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>