dtls: use only one cipher context globally
Instead of having at least 2 cipher contexts per connection, one for
encryption and one for decryption, only use one and initialize it new
for every package to be decrypted or encrypted. In DTLS we have to be
able to decrypt and encrypt every package independently of the other
packages so we do not have to store any state expect for the keys
itself. This should save lots of memory per connection.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>