dtls: allocate handshake struct only when needed
Instead of giving every dtls connection a dedicated handshake struct to
store temporary handshake data in it, make it possible to allocate such
a struct when needed. A dtls connection will allocate such a struct
when a handshake is started and will free it when the handshake
finished. This makes it possible to support e.g. 3 concurrent dtls
connections, but only one handshake at a time without wasting too much
memory.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>