dtls: move epoch to security parameter
The epoch is not stored in the dtls connection any more, but now in the
dtls security parameter. When the epoch was changed there is more than
one valid security parameter and each of them has a own valid epoch.
For sending it now uses the first security parameter and the second
could be null or it could contain some alternative security parameter,
when we are currently in the change process and there still could
arrive an old package we want to receive.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>