crypto: check the length of the pre master secret
For the pre master secret a constant length space is used, but there is
no length check, especially when a psk is used the length of the pre
master secret is depending on the length of the PSK, without checking
the length this could cause memory corruption errors.
Add a length check and return a internal error in case it fails.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>