dtls: stop retransmitting only when new valid message was received
Without this patch the retransmitting was stopped after we received any
package from the other peer, but we should only stop retransmitting if
we received a expect message which is the answer for the last message
we send.
In this case peer A interpreted the retransmission of the first flight
of peer B as a answer to his second flight, which was wrong.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>