Changed configuration process for Contiki.
When invoked with option "--with-contiki" configure used to copy
a hard-coded Makefile and header file instead of generating them
from their according *.in templates. Now, configure performs the
platform checks as specified in configure.in. For platform-specific
adjustments, dtls_config.h can include configuration files that
are given in the subdirectory platform-specific.

To invoke configure for building with Contiki, the host and build
options must be passed, e.g.

  ./configure --with-contiki --build=x86_64-linux-gnu \
     --host=arm-none-eabi

(Seems not to work with msp430-gcc yet.)