get_psk_key may return empty result
The callback function get_psk_key is used at three places:

1. After the server has sent the ServerHello message to determine
   if a ServerKeyExchange message with a psk_hint should be sent.
2. When the client creates the ClientKeyExchange message to select
   the psk_identity.
3. To calculate the key block using the psk as pre-master secret.

In case 1, the function gets called with its parameter id set to zero
to indicate that the keystore may select a suitable identity for that
particular session. In cases where no psk_hint should be sent,
get_psk_key can simply set the result parameter to NULL to suppress
the ServerKeyExchange message.

For cases 2 and 3, setting the result to NULL is an error.