simple key store example with three different identities
get_psk_key() has a static array with pre-configured identities and
keys that are searched when the parameter id is not NULL. Otherwise,
the result will be set to NULL (could be changed to any existing
record to trigger sending the ServerKeyExchange message with a
psk_hint).

When id is set and the identity is not contained in the keystore,
get_psk_key() returns -1, which will tear down the DTLS handshake
with an alert.