optionally read identity and PSK from file
When built with support for pre-shared key mode, the client
provides to new options -i and -k that take a filename as their
argument. The files must contain the psk_identity and the
shared secret, respectively.

When the given file cannot be opened the hard-coded default is
used for backwards compatibility. Change this behaviour for
real-world applications.