Re-introduced combined callback function for PSK material
The callback function(s) for requesting psk_hint, psk_identity
and the psk itself are again combined into a single function.
A new parameter type indicates which of the three is requested.
As output parameter, a simple character buffer is used that
must be filled by the callee. The return value of the function
denotes the number of bytes written to the result buffer on
success, or a value less than zero on error.