Bugfix: error result in get_psk_info() callback
get_psk_info() is called during handshake to select a PSK that
is associated with the given PSK identity. A positive result indicates
the actual number of bytes written to the result parameter.
A negative value indicates an error.

As the codes starting with DTLS_ALERT_ are positive values, they
must be transformed with dtls_alert_fatal_create() before return.

This bugfix addresses ticket #5 "PSK identity is not checked".