Pass psk_identity_hint to call-back function get_psk_info()
When a psk_identity_hint is received from the server, it is
temporarily stored in the keyx parameter block as keyx.psk.identity.
This information now is passed in the parameter id to the
get_psk_info() call-back for type DTLS_PSK_IDENTITY to enable
the client selecting proper credentials.
When the call-back function signals success, the result will be
copied to keyx.psk.identity, hence overriding the psk_identity_hint.

This fix closes ticket #6 "Issue when using PSK Identity Hint".