configure.in: Remove AC_FUNC_MALLOC
The check for malloc usually is not required but can fail when
cross-compiling. In that case, autoconf looks for rpl_malloc()
which would need to be implemented by the application. A common
way of dealing with this is to remove AC_FUNC_MALLOC from
configure and have the user provide a malloc() wrapper if required.

Change-Id: I24ab688d0ffdcf7233db9f250a5d8255d72a9ad4