fixed Makefile variables for cross-compilation
When configured for a different host, the variables CC, AR and
RANLIB have not been set correctly. With this fix, tinydtls can
be compiled for another platform when it is configured as a
submodule of an automake-enabled project.

For example, when the dtls branch of libcoap[1] can be cross-compiled
for the raspberry pi as follows:

  ./configure --with-host=arm-linux-gnueabihf --disable-shared

Note that cross-compilation does not yet work without automake as the
m4 macro AM_PROG_AR depends on automake.

[1] https://github.com/obgm/libcoap/tree/dtls

Change-Id: If23080dc7754a48434e7ac787c0bddda1dd9530b