configure.ac: workaround for autoconf bug
When not using automake, invocation of autoreconf does not install
missing scripts such as install-sh but makes configure check for
an install script. As a workaround, this change creates an empty
file 'install-sh' in the current directory unless it is already
present.

Note that the Makefiles use 'cp' for installation.

Change-Id: Ief05a5672344c2e63f106c4442dc6f954d93983c