When working with limited stacks, having sendbuf on the stack can cause issues
When there is a limited stack size, having the send buffer (which everything
is built into prior to sending) on the stack can easily cause stack overflows.

dtls_mutex.h:

Create dtls_ mutex functions which are platform independent.

dtls.c:

Move sendbuf out of the functions' stack and make it a static variable
if DTLS_CONSTRAINED_STACK is defined.

Protect the usage of sendbuf by mutexes.

Signed-off-by: Jon Shallow <supjps-libcoap@jpshallow.com>