dtls_send_client_key_exchange does not preset all bytes before being sent out
Seen in the client when using using tinydtls client and tinydtls server
with a mismatch of PSK keys. See PR #8 comments for example of valgrind output.

dtls.c:

Clear out the on stack buf variable to zero in dtls_send_client_key_exchange()
before adding in encrypted data (which could be rounded up to a block size)
later on in the function.

Signed-off-by: Jon Shallow <supjps-libcoap@jpshallow.com>