sha256: Handle r and s values that do not use 32 bytes
In ASN.1, these are positive signed integer values, big-endian encoded.

There is an additional 0x00 prefix if the number has the signed bit set, so
that the number remains positive - hence length can be > 32.

If the leading bits are 0, and there is more than 8 of them, then the leading
0x00 byte(s) can be dropped - hence length can be < 32.

This fix allows the dtls_check_ecdsa_signature_elem() to work correctly.
It also corrects dtls_ec_key_from_uint32_asn1() so that the ASN.1 integer is
built correctly.

Signed-off-by: Jon Shallow <supjps-libcoap@jpshallow.com>