Add separate stateless processing for ClientHello of epoch 0.
Use the record and message sequence numbers of that ClientHellos for
either sending a HelloVerifyRequest or starting the handshake on the
server side with sending a ServerHello.
To reduce the complexity, this stateless processing is separated in a
set of special functions, which use the new ephemeral_peer instead of
the old peer. The names of those function follow the pattern xxx_0_yyy
to indicate their usage in epoch 0.

Using a HelloRequest in epoch 0 is removed. If required, a
clarification about the processing details is required ahead.

(Fix some minor typos in comments and log messages.)

Signed-off-by: Achim Kraus <achim.kraus@bosch.io>