dtls.c: fix retransmission of last flight.
The last dtls handshake flights are very special on both sides. The
server doesn't retransmit the last flight on it's own. The server only
retransmits the last flight, if the client's last flight is received
again.
Therefore it's important, that the client doesn't stop the
retransmission of the last flight just by receiving the server's
ChangeCipherSpec.
Fix it by moving the dtls_stop_retransmission into the error handling.

Signed-off-by: Achim Kraus <achim.kraus@bosch.io>