autotools: Generate and install pkg-config file on build
In some cases it might be advantageous to install tinydtls to some
(system-wide) prefix (e.g. `/usr/local`) and use it from there.

Because the tinydtls headers require the header sub-directory (i.e.
`/usr/local/include/tinydtls` if the prefix is `/usr/local`) to be in the
header search path, software that uses tinydtls has to add `-I[HEADER_PATH]`
to its compiler flags (as well as the `-L[LIB_PATH] -ltinydtls` linker flags
if the prefix is not part of the standard library search path).

By providing a pkg-config file, it is possible for build scripts in dependent
projects to automatically set the appropiate compiler and linker flags for
tinydtls without having to manually find the tinydtls subdirectory in one of
its header search path entries.

The format of the pkg-config file follows the guide in
https://people.freedesktop.org/~dbn/pkg-config-guide.html, with the same
adjustments that were made in e.g. libcoap
(https://github.com/obgm/libcoap/blob/develop/libcoap-3.pc.in) to allow for
integration into autotools (e.g. setting the package name, URL and version
based on the values specified in `AC_INIT`).

Signed-off-by: Hugo Hakim Damer <hdamer@uni-bremen.de>