dtls.c: optionally suppress CloseNotify in dtls_destroy_peer()
The parameter unlink for dtls_destroy_peer() has been in use for
several years and therefore can be safely removed.
The third argument for dtls_destroy_peer() now is treated as
flags. The only value currently defined is DTLS_DESTROY_CLOSE to
indicate that the connection should be closed (no CloseNotify is sent
when the session is already closed or closing).

Change-Id: I752708b50d4fbe926802a6b1be44a898acb2d485